Flowchart Maker
The idea for this was an assignment in APCSP of creating a flowchart of an algorithm that uses sequencing, selection and iteration.
Looking at the example, the obvious reached out to me: The flowchart was essentially a directed graph!
Well, why should I draw out a hard-coded flowchart when I can create a scalable program in C++ (Yes, the class is in Java but I dislike it) that builds a virtual flowchart for you, allowing you to traverse a path after initializing your nodes! I couldn't pass off this opportunity for creating this fun and impressive project, so I just had to implement it!
This program uses a combination of adjacency lists and hashmaps to accurately allow one to take a tour through the algorithm!
And not only does this flowchart support sequencing, selection and iteration... it includes all three elements in the code as well!